# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "DIMO Device Definitions API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/compatibility/r1-sheet": {
            "get": {
                "description": "gets r1 MMY compatibility google sheet in JSON form. returns an array of below objects",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device-definitions"
                ],
                "summary": "gets r1 MMY compatibility google sheet in JSON form",
                "operationId": "GetCompatibilityR1Sheet",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_DIMO-Network_device-definitions-api_internal_core_queries.CompatibilitySheetRow"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/device-definitions": {
            "get": {
                "description": "gets a specific device definition by make model and year.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device-definitions"
                ],
                "summary": "gets a specific device definition by make model and year.",
                "operationId": "GetDeviceDefinitionByMMY",
                "parameters": [
                    {
                        "type": "string",
                        "description": "make",
                        "name": "make",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "model",
                        "name": "model",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "year",
                        "name": "year",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_DIMO-Network_device-definitions-api_internal_core_models.GetDeviceDefinitionQueryResult"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/device-definitions/all": {
            "get": {
                "description": "gets a device definition",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device-definitions"
                ],
                "summary": "gets all device definitions by Makes, models, and years",
                "operationId": "GetDeviceDefinitionAll",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/device-definitions/autocomplete": {
            "get": {
                "description": "gets a device definition Autocomplete",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device-definitions"
                ],
                "summary": "gets device definitions autocomplete",
                "operationId": "GetDeviceDefinitionAutocomplete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "query filter",
                        "name": "query",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_DIMO-Network_device-definitions-api_internal_core_queries.GetAllDeviceDefinitionByAutocompleteQueryResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/device-definitions/decode-vin": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "decode a vin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device-definitions"
                ],
                "summary": "returns device definition id corresponding to a given vin",
                "operationId": "DecodeVIN",
                "parameters": [
                    {
                        "description": "Decode VIN request",
                        "name": "decodeRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.DecodeVINRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response with definition ID.",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.DecodeVINResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/device-definitions/search": {
            "get": {
                "description": "gets a device definition by search filter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device-definitions"
                ],
                "summary": "gets device definitions by search filter",
                "operationId": "GetDeviceDefinitionSearch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "query filter",
                        "name": "query",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "make Slug",
                        "name": "makeSlug",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "model Slug",
                        "name": "modelSlug",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "year",
                        "name": "year",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "pageSize",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_DIMO-Network_device-definitions-api_internal_core_queries.GetAllDeviceDefinitionBySearchQueryResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/device-definitions/search-r1": {
            "get": {
                "description": "gets r1 compatibility search by filter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device-definitions"
                ],
                "summary": "gets r1 MMY compatibility by search filter",
                "operationId": "GetR1CompatibilitySearch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "query filter",
                        "name": "query",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "pageSize",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_DIMO-Network_device-definitions-api_internal_core_queries.GetR1CompatibilitySearchQueryResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/device-definitions/{id}": {
            "get": {
                "description": "gets a device definition",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device-definitions"
                ],
                "summary": "gets a device definition",
                "operationId": "GetDeviceDefinitionByID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "device definition id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_DIMO-Network_device-definitions-api_internal_core_models.GetDeviceDefinitionQueryResult"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/v2/device-definitions/{id}": {
            "get": {
                "description": "gets a device definition by the new slug id eg. ford_escape_2021",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device-definitions"
                ],
                "summary": "gets a device definition by the new slug id eg. ford_escape_2021",
                "operationId": "GetDeviceDefinitionByIDv2",
                "parameters": [
                    {
                        "type": "string",
                        "description": "slug definition id eg. ford_escape_2021",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_DIMO-Network_device-definitions-api_internal_infrastructure_gateways.DeviceDefinitionTablelandModel"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_DIMO-Network_device-definitions-api_internal_core_models.DeviceMake": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "externalIdsTyped": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_DIMO-Network_device-definitions-api_internal_core_models.ExternalID"
                    }
                },
                "external_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "hardware_template_id": {
                    "$ref": "#/definitions/null.String"
                },
                "id": {
                    "type": "string"
                },
                "logo_url": {
                    "$ref": "#/definitions/null.String"
                },
                "metadata": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "metadataTyped": {
                    "$ref": "#/definitions/github_com_DIMO-Network_device-definitions-api_internal_core_models.DeviceMakeMetadata"
                },
                "name": {
                    "type": "string"
                },
                "nameSlug": {
                    "type": "string"
                },
                "oem_platform_name": {
                    "$ref": "#/definitions/null.String"
                },
                "tokenId": {
                    "$ref": "#/definitions/big.Int"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_DIMO-Network_device-definitions-api_internal_core_models.DeviceStyle": {
            "type": "object",
            "properties": {
                "deviceDefinitionId": {
                    "type": "string"
                },
                "externalStyleId": {
                    "type": "string"
                },
                "hardware_template_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "metadata": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_DIMO-Network_device-definitions-api_internal_core_models.DeviceTypeAttribute"
                    }
                },
                "name": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "subModel": {
                    "type": "string"
                }
            }
        },
        "github_com_DIMO-Network_device-definitions-api_internal_core_models.DeviceTypeAttribute": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "required": {
                    "type": "boolean"
                },
                "type": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "github_com_DIMO-Network_device-definitions-api_internal_core_models.GetDeviceDefinitionQueryResult": {
            "type": "object",
            "properties": {
                "deviceAttributes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_DIMO-Network_device-definitions-api_internal_core_models.DeviceTypeAttribute"
                    }
                },
                "deviceDefinitionId": {
                    "type": "string"
                },
                "deviceStyles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_DIMO-Network_device-definitions-api_internal_core_models.DeviceStyle"
                    }
                },
                "hardware_template_id": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "make": {
                    "$ref": "#/definitions/github_com_DIMO-Network_device-definitions-api_internal_core_models.DeviceMake"
                },
                "metadata": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "name": {
                    "type": "string"
                },
                "nameSlug": {
                    "type": "string"
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "verified": {
                    "type": "boolean"
                }
            }
        },
        "github_com_DIMO-Network_device-definitions-api_internal_core_queries.CompatibilitySheetRow": {
            "type": "object",
            "properties": {
                "compatible": {
                    "type": "string"
                },
                "definitionId": {
                    "type": "string"
                },
                "make": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "github_com_DIMO-Network_device-definitions-api_internal_core_queries.GetAllDeviceDefinitionAutocompleteItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_DIMO-Network_device-definitions-api_internal_core_queries.GetAllDeviceDefinitionByAutocompleteQueryResult": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_DIMO-Network_device-definitions-api_internal_core_queries.GetAllDeviceDefinitionAutocompleteItem"
                    }
                }
            }
        },
        "github_com_DIMO-Network_device-definitions-api_internal_core_queries.GetAllDeviceDefinitionBySearchQueryResult": {
            "type": "object",
            "properties": {
                "deviceDefinitions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_DIMO-Network_device-definitions-api_internal_core_queries.GetAllDeviceDefinitionItem"
                    }
                },
                "facets": {
                    "$ref": "#/definitions/github_com_DIMO-Network_device-definitions-api_internal_core_queries.GetAllDeviceDefinitionFacet"
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_DIMO-Network_device-definitions-api_internal_core_queries.GetAllDeviceDefinitionPagination"
                }
            }
        },
        "github_com_DIMO-Network_device-definitions-api_internal_core_queries.GetAllDeviceDefinitionFacet": {
            "type": "object",
            "properties": {
                "makes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_DIMO-Network_device-definitions-api_internal_core_queries.GetAllDeviceDefinitionFacetItem"
                    }
                },
                "models": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_DIMO-Network_device-definitions-api_internal_core_queries.GetAllDeviceDefinitionFacetItem"
                    }
                },
                "years": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_DIMO-Network_device-definitions-api_internal_core_queries.GetAllDeviceDefinitionFacetItem"
                    }
                }
            }
        },
...
        "github_com_DIMO-Network_device-definitions-api_internal_core_queries.GetAllDeviceDefinitionItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "legacy_ksuid": {
                    "description": "nolint",
                    "type": "string"
                },
                "make": {
                    "type": "string"
                },
                "model": {
                    "description": "ManufacturerTokenID int    `json:\"manufacturerTokenId\"` // todo",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_DIMO-Network_device-definitions-api_internal_core_queries.GetR1CompatibilitySearchQueryResult": {
            "type": "object",
            "properties": {
                "deviceDefinitions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_DIMO-Network_device-definitions-api_internal_core_queries.GetR1SearchEntryItem"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_DIMO-Network_device-definitions-api_internal_core_queries.GetAllDeviceDefinitionPagination"
                }
            }
        },
...
        "github_com_DIMO-Network_device-definitions-api_internal_infrastructure_gateways.DeviceDefinitionMetadata": {
            "type": "object",
            "properties": {
                "device_attributes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_DIMO-Network_device-definitions-api_internal_infrastructure_gateways.DeviceTypeAttribute"
                    }
                }
            }
        },
        "github_com_DIMO-Network_device-definitions-api_internal_infrastructure_gateways.DeviceDefinitionTablelandModel": {
            "type": "object",
            "properties": {
                "devicetype": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "imageuri": {
                    "type": "string"
                },
                "ksuid": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/github_com_DIMO-Network_device-definitions-api_internal_infrastructure_gateways.DeviceDefinitionMetadata"
                },
                "model": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "internal_api_handlers.DecodeVINRequest": {
            "type": "object",
            "properties": {
                "countryCode": {
                    "description": "3 letter ISO",
                    "type": "string"
                },
                "vin": {
                    "type": "string"
                }
            }
        },
        "internal_api_handlers.DecodeVINResponse": {
            "type": "object",
            "properties": {
                "deviceDefinitionId": {
                    "description": "new name slug based id, can use this to query identity-api",
                    "type": "string"
                },
                "newTransactionHash": {
                    "description": "if a new device definition was created, the tableland transaction hash from the insert statement. Check this has completed before querying the ID",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}